3.3.5

+++ fe/fe-core/src/main/java/com/starrocks/common/profile/TracerImpl.java
@@ -85,7 +85,7 @@ class TracerImpl extends Tracer {
         tracerCost.stop();
     }
 
-    public void count(String name, long count) {
+    public void count(String name, int count) {
         tracerCost.start();
         varTracer.count(timePoint(), name, count);
         tracerCost.stop();
